+++ package.json
@@ -17,7 +17,7 @@
     "@nuxt/components": "^0.3.3",
     "@nuxt/content": "^1.3.1",
     "date-fns": "^2.14.0",
-    "nuxt": "^2.0.0",
+    "nuxt": "^2.12.2",
     "reading-time": "^1.2.0"
   },
   "devDependencies": {
